requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.0
gradio>=4.0.0
google-generativeai>=0.1.0
requests>=2.31.0
python-dotenv>=1.0.0